Updates on language/library features and ABI
Provide a bit more detail on CMake's compile features vs Standard
Library implementation details and how this affects configure
time checks.

Add a section on the potential for ABI incompatibilities between
two C++ standards and how this relates to allow user forcing of the
Standard to compile a project. Note current ideas for making this
transparent as possible to the end user.