Demo workaround for std::make_unique
In certain cases, may want to use C++ Standard Library functionality
for which the current library implementation does not provide, or
which may be only for release in a future standard.

Demo provision of a implementation of std::make_unique for use
in C++11 (or C++14 implementations that do not provide it). Export
variable holding result of test for std::make_unique as a #cmakedefine
and wrap the implementation inside an #ifdef on this.